requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
sqlalchemy>=2.0.0
python-multipart>=0.0.6
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-dotenv>=1.0.0
google-generativeai>=0.3.0
pytesseract>=0.3.10
Pillow>=10.0.0
PyPDF2>=3.0.0
python-docx>=1.1.0
openpyxl>=3.1.0
requests>=2.31.0
aiofiles>=23.2.0
jinja2>=3.1.0
